simplify error handling when writing
It's annoying to check the return value each time we write and it is not
really useful because we cannot recover from this kind of error.

We now signal an error and close the connection.

Note that we close the connection *after* the callback has returned to avoid
signalling a CONN_CLOSED event during the execution of the callback. If it was
to happen, the callback could try to use private data destroyed in the
CONN_CLOSED event handler.