fix muxer when fds are reused
When we unwatch a fd, we mark it as disabled (enabled = false). If we are in
an event callback we do not delete the watcher and wait for the end of the
callback.

If we are watching the same fd just after unwatching it (this happens when
closing the last used fd and immediately creating a new one), in the same
callback, the watcher already exists. It is disabled, and has been deactivated
in the event backend, but is still marked as registered (it should not be).

With this fix, we simply reactivate the watcher.